+++ packages/dev/core/src/Meshes/Node/Blocks/Sources/cylinderBlock.ts
@@ -23,8 +23,8 @@ export class CylinderBlock extends NodeGeometryBlock {
         this.registerInput("diameter", NodeGeometryBlockConnectionPointTypes.Float, true, 1);
         this.registerInput("diameterTop", NodeGeometryBlockConnectionPointTypes.Float, true, 0);
         this.registerInput("diameterBottom", NodeGeometryBlockConnectionPointTypes.Float, true, 0);
-        this.registerInput("subdivisions", NodeGeometryBlockConnectionPointTypes.Float, true, 1);
-        this.registerInput("tessellation", NodeGeometryBlockConnectionPointTypes.Float, true, 24);
+        this.registerInput("subdivisions", NodeGeometryBlockConnectionPointTypes.Int, true, 1);
+        this.registerInput("tessellation", NodeGeometryBlockConnectionPointTypes.Int, true, 24);
         this.registerInput("arc", NodeGeometryBlockConnectionPointTypes.Float, true, 1.0);
 
         this.registerOutput("geometry", NodeGeometryBlockConnectionPointTypes.Geometry);
